Reexport exponentiation from Prelude with exponent type forced to Int
This finally removes the warning about defaulting to Integer (when
using exponent literals), without needing any additional annotation.

+++ Prb/Prb016.hs
@@ -3,7 +3,7 @@ module Prb.Prb016 (prb16) where
 import Data.Char (digitToInt)
 
 bigNum :: Integer
-bigNum = 2^(1000::Int)
+bigNum = 2^1000
 
 prb16 :: IO Int
 prb16 = return $ foldl1 (+) $ map digitToInt $ show $ bigNum
